+++ src/index.ts
@@ -170,6 +170,8 @@ import {
   MatSidenavModule,
 } from '@angular/material';
 
+import { NgxCurrencyModule } from 'ngx-currency';
+
 const components = [
   CardHeaderContainerComponent,
   CardSubtitleComponent,
@@ -241,6 +243,24 @@ const components = [
   MessagesComponent,
 ];
 
+const currencyConfig = {
+  align: 'right',
+  allowNegative: true,
+  allowZero: true,
+  decimal: new CurrencyPipe(navigator.languages ? navigator.languages[0] : navigator.language)
+    .transform(1000, undefined, undefined, '4.1-1')
+    .split('000')[0]
+    .substr(-1, 1),
+  precision: 2,
+  prefix: '',
+  suffix: '',
+  thousands: new CurrencyPipe(navigator.languages ? navigator.languages[0] : navigator.language)
+    .transform(1000, undefined, undefined, '4.1-1')
+    .split('000')[1]
+    .substr(0, 1),
+  nullable: false,
+};
+
 @NgModule({
   imports: [
     CommonModule,
@@ -265,6 +285,7 @@ const components = [
     TranslationModule,
     MatSidenavModule,
     LoggerModule,
+    NgxCurrencyModule.forRoot(currencyConfig),
   ],
   declarations: components,
   exports: [
@@ -273,6 +294,7 @@ const components = [
     ExpandableFabItemComponent,
     TextListItemComponent,
     DataTableContainerComponent,
+    NgxCurrencyModule,
   ],
   providers: [TranslationService, CurrencyPipe],
 })